Integrate startup service, fix many bugs and add javadoc comment
Now the startup phase works correctly. All function is designed for build1 presentation.
