--- mapEditor/Source Codes/Order.java
@@ -1,5 +1,0 @@
-package comp691.project.map;
-
-abstract class Order {
-    abstract void execute();
-}
