+++ unusedcode/Source Codes/Order.java
@@ -1,0 +1,5 @@
+package comp691.project.map;
+
+abstract class Order {
+    abstract void execute();
+}
